Managers can view all employees
Managers can now view all of the company's employees. Employee and
manager webpages consist of buttons that will be used to test some
functionality to be added in the future.

I more concerned with function over form at the moment, so I intend on
focusing on user stories. Proper styling will be added here and there as
the project progresses.